refactor: rename all api models to dtos, rename packages and apply dto mapping in the same file
