+++ app/src/main/java/com/monstarlab/features/auth/data/api/AuthApi.kt
@@ -1,0 +1,14 @@
+package com.monstarlab.features.auth.data.api
+
+import com.monstarlab.features.auth.data.api.dtos.TokenResponseDTO
+import retrofit2.Response
+import retrofit2.http.Field
+import retrofit2.http.FormUrlEncoded
+import retrofit2.http.POST
+
+interface AuthApi {
+
+    @FormUrlEncoded
+    @POST("login")
+    suspend fun postLogin(@Field("email") email: String, @Field("password") password: String): Response<TokenResponseDTO>
+}
