--- app/src/main/java/com/monstarlab/features/auth/domain/AuthToken.kt
@@ -1,3 +1,0 @@
-package com.monstarlab.features.auth.domain
-
-data class AuthToken(val token: String)
