+++ app/src/main/java/com/monstarlab/features/auth/domain/models/AuthToken.kt
@@ -1,0 +1,3 @@
+package com.monstarlab.features.auth.domain.models
+
+data class AuthToken(val token: String)
