--- app/src/main/java/com/monstarlab/features/resources/data/ResourceDto.kt
@@ -1,14 +1,0 @@
-package com.monstarlab.features.resources.data
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class ResourceDto(
-    val id: Int,
-    val name: String,
-    val year: Int,
-    val color: String,
-    @SerialName("pantone_value")
-    val pantoneValue: String
-)
