--- app/src/main/java/com/monstarlab/features/resources/data/ResourceMapping.kt
@@ -1,13 +1,0 @@
-package com.monstarlab.features.resources.data
-
-import com.monstarlab.features.resources.domain.Resource
-
-fun ResourceDto.toEntity(): Resource {
-    return Resource(
-        id = id,
-        name = name,
-        year = year,
-        color = color,
-        pantoneValue = pantoneValue
-    )
-}
