+++ app/src/main/java/com/monstarlab/features/resources/data/api/ResourcesApi.kt
@@ -1,0 +1,11 @@
+package com.monstarlab.features.resources.data.api
+
+import com.monstarlab.features.resources.data.api.dtos.ResourcesResponseDTO
+import retrofit2.Response
+import retrofit2.http.GET
+import retrofit2.http.Query
+
+interface ResourcesApi {
+    @GET("unknown")
+    suspend fun getResources(@Query("page") page: Int = 1): Response<ResourcesResponseDTO>
+}
