+++ app/src/main/java/com/monstarlab/features/resources/data/api/dtos/ResourcesResponseDTO.kt
@@ -1,15 +1,15 @@
-package com.monstarlab.features.resources.data
+package com.monstarlab.features.resources.data.api.dtos
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class ResourcesResponse(
+data class ResourcesResponseDTO(
     val page: Int,
     @SerialName("per_page")
     val perPage: Int,
     val total: Int,
     @SerialName("total_pages")
     val totalPages: Int,
-    val data: List<ResourceDto>
+    val data: List<ResourceDTO>
 )
