+++ app/src/main/java/com/monstarlab/features/user/data/api/dtos/UserResponseDTO.kt
@@ -1,0 +1,8 @@
+package com.monstarlab.features.user.data.api.dtos
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class UserResponseDTO(
+    val data: UserDTO
+)
