--- app/src/main/java/com/monstarlab/features/user/data/models/UserDto.kt
@@ -1,15 +1,0 @@
-package com.monstarlab.features.user.data.models
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class UserDto(
-    val id: Int,
-    val email: String,
-    @SerialName("first_name")
-    val firstName: String,
-    @SerialName("last_name")
-    val lastName: String,
-    val avatar: String
-)
