--- app/src/main/java/com/monstarlab/features/user/data/models/UserMapping.kt
@@ -1,12 +1,0 @@
-package com.monstarlab.features.user.data.models
-
-import com.monstarlab.features.user.domain.User
-
-fun UserDto.toUser(): User {
-    return User(
-        email = email,
-        firstName = firstName,
-        lastName = lastName,
-        avatar = avatar
-    )
-}
