--- app/src/main/java/com/monstarlab/features/user/data/models/UserResponse.kt
@@ -1,8 +1,0 @@
-package com.monstarlab.features.user.data.models
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class UserResponse(
-    val data: UserDto
-)
