+++ app/src/main/java/com/monstarlab/features/user/domain/UserRepository.kt
@@ -2,9 +2,9 @@ package com.monstarlab.features.user.domain
 
 import com.monstarlab.core.extensions.mapSuccess
 import com.monstarlab.core.persistence.Repository
-import com.monstarlab.features.user.data.UserPreferenceStore
-import com.monstarlab.features.user.data.UsersApi
-import com.monstarlab.features.user.data.models.toUser
+import com.monstarlab.features.user.data.storage.UserPreferenceStore
+import com.monstarlab.features.user.data.api.UsersApi
+import com.monstarlab.features.user.data.api.dtos.toUser
 import javax.inject.Inject
 
 class UserRepository @Inject constructor(
